Respect geth structure due work, optimised sql query to db
